+++ src/i18n/locales/en/translations.json
@@ -127,6 +127,7 @@
   "liquidity.asset.positions.position.currentValue": "Current value",
   "liquidity.asset.positions.position.initialValue": "Initial value",
   "liquidity.asset.positions.position.title": "Liquidity position #{{ index }}",
+  "liquidity.asset.positions.position.tooltip": "Sum of underlying asset and LRNA: $t(value.token)",
   "liquidity.asset.positions.position.shares": "{{ shares, bignumber(type: 'token') }} asset shares",
   "liquidity.asset.capacity": "<0>Estimated asset cap:</0> <1>{{ filled, compact }} / {{ capacity, compact }} {{ symbol }}</1>",
   "liquidity.asset.capacity.full": "<0>Asset cap reached:</0> <1>{{ filled, compact }} / {{ capacity, compact }} {{ symbol }}</1>",
