+++ src/sections/pools/pool/details/PoolValue.tsx
@@ -7,7 +7,7 @@ import Skeleton from "react-loading-skeleton"
 import { OmnipoolPool } from "sections/pools/PoolsPage.utils"
 import { usePoolDetailsTradeVolume } from "sections/pools/pool/details/PoolDetails.utils"
 import { BN_NAN } from "utils/constants"
-import { SInfoIcon } from "./PoolValue.styled"
+import { SInfoIcon } from "../Pool.styled"
 
 type PoolValueProps = { pool: OmnipoolPool; className?: string }
